+++ packages/example-react/project/app/Response/props.js
@@ -1,0 +1,10 @@
+import PropTypes from 'prop-types';
+
+export default {
+  stateStore: PropTypes.shape({
+    form: PropTypes.any
+  }),
+  appStore: PropTypes.shape({
+    state: PropTypes.any
+  })
+};