+++ packages/workflow-plugin-angular/test.js
@@ -1,10 +1,19 @@
+/* eslint-disable jest/no-jest-import */
+/* eslint-disable global-require */
+/* eslint-disable import/no-dynamic-require */
 // https://github.com/facebookincubator/create-react-app/blob/master/packages/react-scripts/utils/createJestConfig.js
+const { existsSync } = require('fs');
+
+// https://github.com/facebook/jest/issues/7704#issuecomment-458552963
+require('jest/node_modules/jest-cli/build/cli');
+
 const jest = require('jest');
 const settings = require('@availity/workflow-settings');
 const path = require('path');
 
 function create() {
   const rootDir = settings.project();
+  const jestInitExists = existsSync(`${path.join(settings.app(), 'jest.init.js')}`);
 
   const config = {
     collectCoverageFrom: ['project/app/**/*.js'],
@@ -16,7 +25,12 @@ function create() {
       '^.+\\.css$': `${require.resolve('./jest/css.js')}`,
       '^(?!.*\\.(js|css|json)$)': `${require.resolve('./jest/file.js')}`
     },
-    setupTestFrameworkScriptFile: `${require.resolve(path.join(settings.app(), 'jest.init.js'))}`,
+    setupFilesAfterEnv: jestInitExists
+    ? require(path.join(settings.app(), 'jest.init.js'))
+    : [
+      'angular',
+      'angular-mocks'
+    ],
     transformIgnorePatterns: ['[/\\\\]node_modules[/\\\\](?!@?av).+\\.(js|jsx|html)$'],
     moduleDirectories: ['node_modules', 'project/app', 'app'],
     testMatch: [