+++ packages/workflow-settings/webpack/index.js
@@ -1,3 +1,4 @@
+/* eslint-disable global-require */
 module.exports = {
   css: require('./loader-css'),
   less: require('./loader-less'),
@@ -5,5 +6,6 @@ module.exports = {
   postcss: require('./loader-postcss'),
   fonts: require('./rule-fonts'),
   images: require('./rule-images'),
+  eslint: require('./loader-eslint'),
   MiniCssExtractPlugin: require('mini-css-extract-plugin')
 };