+++ packages/workflow-settings/webpack/loader-eslint.js
@@ -1,0 +1,11 @@
+module.exports = {
+    loader: "eslint-loader",
+    exclude: /node_modules/,
+    enforce: "pre",
+    test: /\.js?$/,
+    options: {
+        baseConfig: {
+            extends: 'eslint-config-availity'
+        }
+    }
+};