+++ packages/workflow/scripts/plugin.js
@@ -1,10 +1,15 @@
+/* eslint-disable import/no-dynamic-require */
+/* eslint-disable global-require */
 const settings = require('@availity/workflow-settings');
 const Logger = require('@availity/workflow-logger');
 const figures = require('figures');
 
 function plugin(path) {
   let plugin = settings.pkg().availityWorkflow && settings.pkg().availityWorkflow.plugin;
-  plugin = plugin || settings.pkg().devDependencies && Object.keys(settings.pkg().devDependencies).filter(p => /@availity\/workflow-plugin-.+/.test(p));
+  plugin =
+    plugin ||
+    (settings.pkg().devDependencies &&
+      Object.keys(settings.pkg().devDependencies).filter(p => /@availity\/workflow-plugin-.+/.test(p)));
 
   if (!plugin) {
     Logger.failed(`Project must be configured to use React or Angular
@@ -27,16 +32,29 @@ ${figures.pointer}
   }
 
   let file;
+  let err;
 
   try {
     const filePath = `${plugin}/${path}`;
-    // eslint-disable-next-line
     file = require(filePath);
-  } catch (err) {
+  } catch (error) {
+    err = error;
+  }
+
+  if (!file) {
     // Workaround when Lerna linked modules
     // eslint-disable-next-line global-require
     const relative = require('require-relative');
-    file = relative(`${plugin}/${path}`, settings.project());
+    try {
+      file = relative(`${plugin}/${path}`, settings.project());
+    } catch (error) {
+      err = error;
+    }
+  }
+
+  if (!file && err) {
+    Logger.error(err);
+    throw err;
   }
 
   return file;